+++ advancedCG/src/Object.cpp
@@ -38,7 +38,7 @@ bool Object::setMaterial(const Material& mat) {
 
 /* class methods */
 
-float Object::rayHitPosition(const Ray3& ray) const { return -1; }
+double Object::rayHitPosition(const Ray3& ray) const { return -1; }
 Vector3 Object::getNormal(const Vector3& hitPos) const {
     Vector3 empty(0, 0, 0);
     return empty;
